refactor: Replace calculations from double to BigDecimal type for improved precision.

+++ src/main/java/org/example/first/ExchangeDTO.java
@@ -4,13 +4,15 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @JsonPropertyOrder({"baseCurrency", "targetCurrency", "rate", "amount", "convertedAmount"})
 @Data
 @AllArgsConstructor
 public class ExchangeDTO {
     private CurrencyDTO baseCurrency;
     private CurrencyDTO targetCurrency;
-    private double rate;
-    private double amount;
-    private double convertedAmount;
+    private BigDecimal rate;
+    private BigDecimal amount;
+    private BigDecimal convertedAmount;
 }
